requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.4.4
flask-pymongo==4.0.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1
werkzeug==3.0.1
pytest==7.4.4
pytest-cov==4.1.0
flask-swagger-ui==4.14.0
gunicorn
pymongo[srv]==21.2.0
